% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grid.R
\name{add_grid}
\alias{add_grid}
\title{Add Gridded Data}
\usage{
add_grid(x_grid, x_dt)
}
\arguments{
\item{x_grid}{gridded environmental data of class raster}

\item{x_dt}{trawl data.table}
}
\description{
Added a gridded environmental data set to a data.table of clean trawl survey data
}

