% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatStrat.R
\name{calcarea}
\alias{calcarea}
\title{Calculate Area}
\usage{
calcarea(lonlat)
}
\arguments{
\item{lonlat}{a data.frame with numeric longitude in first column, latitude in second}
}
\description{
Calculate the area of a region defined by a vector of lon-lat coordinates
}

