% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_strat.R
\name{check_strat}
\alias{check_strat}
\title{Check Strata}
\usage{
check_strat(X, reg = c("ai", "ebs", "gmex", "goa", "neus", "newf", "sa",
  "sgulf", "shelf", "wcann", "wctri"), append_keep_strat = FALSE,
  prompt_strat_tol = FALSE, strat_tol, plot = TRUE)
}
\arguments{
\item{X}{A trawl data.table}

\item{reg}{region name}

\item{append_keep_strat}{Logical, whether to add the \code{keep_strat} column to \code{X}}

\item{prompt_strat_tol}{Logical, if in \code{\link{interactive}} mode, prompt user for tolerance? If not, and if \code{append_keep_strat} is TRUE and \code{strat_tol} is left \code{\link{missing}}, then a default will be selected for \code{strat_tol}}

\item{strat_tol}{The maximum number of unsampled years that is tolerated for any stratum before all rows corresponding to that stratum have their value in the "keep_strat" column set to FALSE}

\item{plot}{Logical, visualize strata over time and the number of strata sampled in all but N years?}
}
\value{
Nothing. However, potentially modifies X, and may create a figure.
}
\description{
Visualize which strata were visited in each year, and how many strata were visited in all but N years
}
\details{
The aim of the function is to guide the selection of which strata to exclude from analysis because they are not sampled often enough. Having fewer gaps in your data set is better, but sometimes tolerating a tiny amount of missingness can result in huge increases in data; the visualization provided by this funciton will help gauge that tradeoff.
}
\examples{
# quick example for goa
check_strat(clean.goa[,c("year"):=list(as.integer(year))], "goa")

# a more thorough data processing example for shelf
\dontrun{
	# trim shelf
	shelf <- trawlTrim("shelf", c.add=c("val.src", "flag"))
	shelf <- shelf[
		(taxLvl=="species" |taxLvl=="subspecies") & 
		(flag!="bad" | is.na(flag)) & 
		(val.src!="m3" | (!is.na(flag) & flag!="bad"))
	]

	# aggregate species within a haul (among individuals)
	# this means taking the sum of many bio metrics
	shelf <- trawlAgg(
		X=shelf,
		bioFun=sumna,
		envFun=meanna,
		bio_lvl="spp", space_lvl="haulid", time_lvl="haulid",
		bioCols=c("wtcpue"),
		envCols=c("btemp"),
		metaCols=c("reg","common","year","datetime","stratum", "lon", "lat"),
		meta.action=c("unique1")
	)

	# aggregate within a species within stratum
	# refer to the time_lvl column from previous trawlAgg()
	# can use mean for both bio and env
	shelf[,stratum:=ll2strat(lon, lat)]
	shelf <- trawlAgg(
		X=shelf,
		FUN=meanna,
		bio_lvl="spp", space_lvl="stratum", time_lvl="year",
		bioCols=c("wtcpue"),
		envCols=c("btemp"),
		metaCols=c("reg","common", "lon", "lat"),
		meta.action=c("FUN"),
		metaFun=list(reg=unique, common=unique, lon=mean, lat=mean)
	)
	setnames(shelf, "time_lvl", "year")
	shelf[,year:=as.integer(as.character(year))]
	setcolorder(shelf, c(
		"reg", "year", "stratum", "lon", "lat", 
		"spp", "common", "btemp", "wtcpue", "nAgg"
	))
	setkey(shelf, reg, year, stratum, spp, common)
}

}

