% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.ai}
\alias{clean.ai}
\title{Clean Aleutian Islands}
\format{A data.table}
\usage{
clean.ai
}
\description{
Clean data set for the Aleutian Islands bottom trawl survey
}
\keyword{datasets}

