% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.columns.R
\name{clean.columns}
\alias{clean.columns}
\title{Clean Columns}
\usage{
clean.columns(X, reg = c("ai", "ebs", "gmex", "goa", "neus", "newf", "ngulf",
  "sa", "sgulf", "shelf", "wcann", "wctri"))
}
\arguments{
\item{X}{a \code{data.table} containing trawl data}

\item{reg}{a region; should be one of the following: \code{c("ai","ebs","gmex","goa", "neus", "newf", "sa", "sgulf", "shelf", "wcann", "wctri")}}
}
\description{
Clean by ensuring the presence of standard columns.
}
\details{
Add missing columns if they don't exist. These are standard values like \code{wtcpue}, \code{spp}, \code{datetime}, and \code{reg}. I will make a full list in future versions, when the decisions are finalized.
}

