% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.ebs}
\alias{clean.ebs}
\title{Clean Eastern Berring Sea
Clean data set for the Eastern Berring Sea bottom trawl survey}
\format{A dim = 380027 x 47 data.table data.frame:
\tabular{rlll}{
[,1] \tab ref \tab character \tab survey region\cr
[,2] \tab stratum \tab character \tab year of haul\cr
[,3] \tab lat \tab numeric \tab the day and time of the haul\cr
[,4] \tab lon \tab numeric \tab reference taxonomic ID from raw data\cr
[,5] \tab station \tab character \tab species scientific name; Genus species\cr
[,6] \tab year \tab character \tab the common name of the organism sampled\cr
[,7] \tab datetime \tab c("POSIXct", "POSIXt") \tab the genus of the species\cr
[,8] \tab wtcpue \tab numeric \tab the species name of the species\cr
[,9] \tab cntcpue \tab numeric \tab species identification number\cr
[,10] \tab SID \tab integer \tab vessel ID\cr
[,11] \tab depth \tab numeric \tab cruise ID\cr
[,12] \tab btemp \tab numeric \tab the integer haul number within a cruise\cr
[,13] \tab stemp \tab numeric \tab a unique identifier for the haul; vessel ID - cruise ID - haul number\cr
[,14] \tab vessel \tab character \tab the statistical stratum of the haul\cr
[,15] \tab cruise \tab character \tab the station ID for the haul\cr
[,16] \tab haul \tab character \tab the area of the statistical stratum (km2)\cr
[,17] \tab stratumarea \tab numeric \tab latitude of the haul\cr
[,18] \tab haulid \tab character \tab longitude of the haul, in western hemisphere degrees (for lon > 0, do lon-360)\cr
[,19] \tab season \tab character \tab insert_description_here\cr
[,20] \tab reg \tab character \tab the maximum depth of the water at the location of the haul\cr
[,21] \tab val.src \tab character \tab insert_description_here\cr
[,22] \tab tbl.row \tab integer \tab insert_description_here\cr
[,23] \tab mtch.src \tab integer \tab insert_description_here\cr
[,24] \tab tax.src \tab character \tab insert_description_here\cr
[,25] \tab spp \tab character \tab water temperature at the surface at the location of the haul\cr
[,26] \tab common \tab character \tab water temperature at the bottom at the location of the haul\cr
[,27] \tab taxLvl \tab character \tab insert_description_here\cr
[,28] \tab species \tab character \tab weight (mass) of the catch\cr
[,29] \tab genus \tab character \tab number of individuals caught per hectare in the haul\cr
[,30] \tab family \tab character \tab insert_description_here\cr
[,31] \tab order \tab character \tab insert_description_here\cr
[,32] \tab class \tab character \tab insert_description_here\cr
[,33] \tab superclass \tab character \tab insert_description_here\cr
[,34] \tab subphylum \tab character \tab insert_description_here\cr
[,35] \tab phylum \tab character \tab insert_description_here\cr
[,36] \tab kingdom \tab character \tab insert_description_here\cr
[,37] \tab trophicDiet \tab character \tab insert_description_here\cr
[,38] \tab trophicOrig \tab character \tab insert_description_here\cr
[,39] \tab Picture \tab character \tab insert_description_here\cr
[,40] \tab trophicLevel \tab numeric \tab insert_description_here\cr
[,41] \tab trophicLevel.se \tab numeric \tab insert_description_here\cr
[,42] \tab tax.src2 \tab character \tab insert_description_here\cr
[,43] \tab conflict \tab logical \tab insert_description_here\cr
[,44] \tab flag \tab character \tab insert_description_here\cr
[,45] \tab website \tab character \tab insert_description_here\cr
[,46] \tab website2 \tab character \tab insert_description_here\cr
[,47] \tab keep.row \tab logical \tab insert_description_here\cr
}}
\usage{
clean.ebs
}
\description{
Clean Eastern Berring Sea
Clean data set for the Eastern Berring Sea bottom trawl survey
}
\keyword{datasets}

