% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.format.R
\name{clean.format}
\alias{clean.format}
\title{Clean Format}
\usage{
clean.format(X, reg = c("ai", "ebs", "gmex", "goa", "neus", "newf", "ngulf",
  "sa", "sgulf", "shelf", "wcann", "wctri"))
}
\arguments{
\item{X}{a \code{data.table} containing trawl data}

\item{reg}{a region; should be one of the following: \code{c("ai","ebs","gmex","goa", "neus", "newf", "sa", "sgulf", "shelf", "wcann", "wctri")}}
}
\description{
Clean by fixing data entry errors, and standardizing value formats
}
\details{
It is this function that makes specific corrections for data entry errors. For example, in one region a tow duration of 3 should have been 30. In another region some of the \code{effort} values were entered as \code{0} or \code{NA}, but should have had a particular value.
This function also ensures that longitude and latitude are in the same format among regions.
Other data entry errors or necessary corrections are implemented here, too.
Dates are not thoroughly formatted here, except in some cases where getting a \code{year}, e.g., requires parsing values out of other columns. POSIX class dates not created.
}

