% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.goa}
\alias{clean.goa}
\title{Clean Gulf of Alaska}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	197067 obs. of  41 variables:
 $ ref            : chr  "" "" "" "" ...
 $ stratum        : chr  "10" "10" "10" "10" ...
 $ lat            : num  54 53.2 53.3 53.6 53.7 ...
 $ lon            : num  -166 -168 -167 -166 -166 ...
 $ station        : chr  "51-40" "25-21" "35-24" "52-33" ...
 $ year           : chr  "1984" "1987" "1987" "1987" ...
 $ datetime       : POSIXct, format: "1984-07-13 17:00:00" "1987-08-27 21:00:00" ...
 $ wtcpue         : num  0.029 0.0835 0.1805 7.2949 0.1837 ...
 $ cntcpue        : num  NA 0.0836 0.173 8.9645 0.1928 ...
 $ SID            : int  68000 30150 30150 30150 30150 68000 71700 30150 71700 30150 ...
 $ depth          : num  75 75 93 84 80 93 82 77 91 90 ...
 $ btemp          : num  6.3 NA 7 NA NA NA NA NA NA NA ...
 $ stemp          : num  6.8 10.4 11.8 5.7 5.7 6.9 NA NA 6.3 NA ...
 $ vessel         : chr  "57" "560" "560" "560" ...
 $ cruise         : chr  "198401" "198701" "198701" "198701" ...
 $ haul           : chr  "107" "444" "441" "15" ...
 $ stratumarea    : num  8333 8333 8333 8333 8333 ...
 $ haulid         : chr  " 57-198401-107" "560-198701-444" "560-198701-441" "560-198701- 15" ...
 $ season         : chr  "summer" "summer" "summer" "spring" ...
 $ reg            : chr  "goa" "goa" "goa" "goa" ...
 $ val.src        : chr  NA NA NA NA ...
 $ tbl.row        : num  NA NA NA NA NA NA NA NA NA NA ...
 $ mtch.src       : num  NA NA NA NA NA NA NA NA NA NA ...
 $ spp            : chr  NA NA NA NA ...
 $ common         : chr  NA NA NA NA ...
 $ taxLvl         : chr  NA NA NA NA ...
 $ species        : chr  NA NA NA NA ...
 $ genus          : chr  NA NA NA NA ...
 $ family         : chr  NA NA NA NA ...
 $ order          : chr  NA NA NA NA ...
 $ class          : chr  NA NA NA NA ...
 $ superclass     : chr  NA NA NA NA ...
 $ subphylum      : chr  NA NA NA NA ...
 $ phylum         : chr  NA NA NA NA ...
 $ kingdom        : chr  NA NA NA NA ...
 $ trophicDiet    : chr  NA NA NA NA ...
 $ trophicOrig    : chr  NA NA NA NA ...
 $ Picture        : chr  NA NA NA NA ...
 $ trophicLevel   : num  NA NA NA NA NA NA NA NA NA NA ...
 $ trophicLevel.se: num  NA NA NA NA NA NA NA NA NA NA ...
 $ keep.row       : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
clean.goa
}
\description{
Clean data set for the Gulf of Alaska bottom trawl survey
}
\keyword{datasets}

