% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.names.R
\name{clean.names}
\alias{clean.names}
\title{Clean Names}
\usage{
clean.names(X, reg = c("ai", "ebs", "gmex", "goa", "neus", "newf", "ngulf",
  "sa", "sgulf", "shelf", "wcann", "wctri"))
}
\description{
Clean by standardizing the column names
}
\details{
template X_reg

Regions tend to have very different column names for what are essentiallythe same measurements, descriptors, etc. This function tries to give everything a standardized name when it's appropriate.
}

