% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.sa}
\alias{clean.sa}
\title{Clean South Atlantic}
\format{A data.table}
\usage{
clean.sa
}
\description{
Clean data set for the US Southeast bottom trawl survey
}
\keyword{datasets}

