% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.tax.R
\name{clean.tax}
\alias{clean.tax}
\title{Clean Taxonomy}
\usage{
clean.tax(X, reg = c("ai", "ebs", "gmex", "goa", "neus", "newf", "ngulf",
  "sa", "sgulf", "shelf", "wcann", "wctri"))
}
\arguments{
\item{X}{a \code{data.table} containing trawl data}

\item{reg}{a region; should be one of the following: \code{c("ai","ebs","gmex","goa", "neus", "newf", "sa", "sgulf", "shelf", "wcann", "wctri")}}
}
\description{
Clean by fixing/ standardizing taxonomic names
}
\details{
relies on spp.key data file that comes with this package. In "/inst/extdata" there is also a .csv version of this file that is being used. Taxonomy corrections are an ongoing process.  
Also adds taxonomic classification, some ecological information (trophic level), common names, whether or not a picture is available with this package, and other miscellanious information about how the taxonomic correction was made.

The \code{ref} column in the output is the name of the original species name/ taxonomic identifier.
}
\seealso{
\code{\link{clean.names}}, \code{\link{clean.format}}, \code{\link{clean.columns}}, \code{\link{clean.tax}} \code{\link{clean.trimRow}}
}

