% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.trimRow.R
\name{clean.trimRow}
\alias{clean.trimRow}
\title{Clean by Indicating Rows to Trim}
\usage{
clean.trimRow(X, reg = c("ai", "ebs", "gmex", "goa", "neus", "newf", "ngulf",
  "sa", "sgulf", "shelf", "wcann", "wctri"))
}
\arguments{
\item{X}{a \code{data.table} containing trawl data}

\item{reg}{a region; should be one of the following: \code{c("ai","ebs","gmex","goa", "neus", "newf", "sa", "sgulf", "shelf", "wcann", "wctri")}}
}
\description{
Cleans by indicating which rows should be dropped
}
\details{
Recommended rows to drop according to Malin's original scripts and what's in the OceanAdapt repo. Rows are not actually dropped; rather, a column called \code{keep.row} is added to the data.table; when \code{keep.row} is \code{FALSE}, it is recommended that the row be dropped.
}

