% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.wctri}
\alias{clean.wctri}
\title{Clean West Coast Triennial}
\format{A data.table}
\usage{
clean.wctri
}
\description{
Clean data set for the West Coast Triennial bottom trawl survey
}
\keyword{datasets}

