% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.spp.key_manual.R
\name{create.spp.key_manual}
\alias{create.spp.key_manual}
\title{Manual entries to species key
Manually add or correct species names, taxonomic, ecological, or reference information for spp.key}
\usage{
create.spp.key_manual(spp.key)
}
\description{
Manual entries to species key
Manually add or correct species names, taxonomic, ecological, or reference information for spp.key
}

