% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatValue.R
\name{cull}
\alias{cull}
\alias{cullExSpace}
\alias{cullParen}
\alias{cullPost2}
\alias{cullSp}
\alias{fixCase}
\title{Cull Bad Species Characters}
\usage{
cull(x)

fixCase(x)

cullExSpace(x)

cullSp(x)

cullParen(x)

cullPost2(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
a character vector that has been altered by removing content unlikely to belong to a species name.
}
\description{
Remove/ fix characters that are unlikely to be part of a species name
}
\details{
\code{cull} performs the following corrections, in this order:  
1. remove extra spaces
2. change to sentence case
3. remove generic species indicators (spp or Sp., e.g.)
4. Remove everything in parentheses
5. Remove words after the first 2
}
\section{Functions}{
\itemize{
\item \code{fixCase}: Fix case to sentence case

\item \code{cullExSpace}: Remove extra spaces

\item \code{cullSp}: Remove generic species indicator

\item \code{cullParen}: Remove parentheses and contents

\item \code{cullPost2}: Remove words after the last two
}}

