% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperStat.R
\name{fill.mean}
\alias{fill.mean}
\title{Fill Mean}
\usage{
fill.mean(x)
}
\arguments{
\item{x}{numeric vector whose NA values to be replaced by the non-NA mean}
}
\description{
Fill the NA values of a vector with the mean of the non-NA portion
}
\seealso{
\code{\link{sumna}}, \code{\link{meanna}}, \code{\link{wtAvg}}, \code{\link{fill.mean}}
}

