% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperFile.R
\name{fread.fwf}
\alias{fread.fwf}
\title{FWF Fread}
\usage{
fread.fwf(..., cols, column_types, column_names)
}
\arguments{
\item{column_types}{character vector indicating classes of columns}

\item{column_names}{column_names character vector of column headers}

\item{\code{...}}{file names to open}

\item{\code{cols}}{vector of integer column widths}
}
\value{
a data.table
}
\description{
Fast read for fixed width format files
}
\details{
uses data.table and LaF packages. The read is performed entirely by \code{\link{laf_open_fwf}}, but the output is converted to a data.table.
}

