% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperMisc.R
\name{getSeason}
\alias{getSeason}
\title{Get Season}
\source{
\url{http://stackoverflow.com/a/9501225/2343633}
}
\usage{
getSeason(DATES)
}
\arguments{
\item{DATES}{dates formatted as Ymd}
}
\value{
a character vector of seasons
}
\description{
Get the season from the date
}
\details{
Get the season of the year, based on solstices and equinoxes.
}
\seealso{
\code{\link{strptime}} for date formatting details, \code{\link{clean.columns}} for where this function is implemented.
}

