% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tax.getSpp.R
\name{getSpp}
\alias{countN}
\alias{getSpp}
\alias{grb.spp1}
\title{Get Species
Lookup, and correct, species names}
\usage{
getSpp(uspp)

countN(x)

grb.spp1(x)
}
\arguments{
\item{uspp:}{character vector of unique species names to be checked}
}
\value{
A data.table with 2 columns; "spp" column contains unchecked species names, "sppCorr" contains corrected (checked) species names
}
\description{
Get Species
Lookup, and correct, species names
}
\details{
Lookup species names using taxize, with option to only lookup names not in a reference data.table of previously-checked names. Currently looks up 1 species name at a time, but function could be modified to look up chunks. Relies heavily on taxize.
}
\section{Functions}{
\itemize{
\item \code{countN}: Count the number of N's in a word

\item \code{grb.spp1}: Grab Species (helper function)
}}
\seealso{
\code{\link{getCmmn}}, \code{\link{getTax}}, \code{\link{getTL}}, \code{\link{updateDB}}
}

