% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{getSppData}
\alias{getSppData}
\title{Spp Names
Scientific names matched to raw taxonomic name entries}
\format{A dim = 6016 x 2 data.table data.frame:
\tabular{rlll}{
[,1] \tab spp \tab character \tab insert_description_here\cr
[,2] \tab sppCorr \tab character \tab insert_description_here\cr
}}
\usage{
getSppData
}
\description{
Spp Names
Scientific names matched to raw taxonomic name entries
}
\seealso{
\code{\link{getCmmn}} \code{\link{getSpp}} \code{\link{getTax}}
}
\keyword{datasets}

