% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tax.getTax.R
\name{getTax}
\alias{getTax}
\title{Get Taxonomy}
\usage{
getTax(sppCorr, class.names = c("species", "genus", "family", "order",
  "class", "superclass", "subphylum", "phylum", "kingdom"))
}
\arguments{
\item{sppCorr}{a scientific name; character vector}

\item{class.names}{the taxonomic classification names to search for and return}
}
\value{
A data.table with columns corresponding to the taxonomic level, the original species name, and \code{taxLvl} as the most specific (lowest) taxonomic level for which a match was found.
}
\description{
Get the taxonomic classification for a Latin name
}
\details{
Uses taxize. Tries several data bases
}
\examples{
getTax(c(
	"Chionoecetes tanneri"
))
}
\seealso{
\code{\link{getCmmn}}, \code{\link{getTax}}
}

