% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HadISST}
\alias{HadISST}
\alias{sst}
\alias{sst.ann}
\alias{sst.mu}
\title{HadISST Surface Temperatures}
\format{raster brick}
\usage{
sst

sst.ann

sst.mu
}
\description{
Surface temperatures from the HadISST data source. Includes terrestrial and aquatic. Gridded.
}
\details{
The full data set is in \code{sst}; \code{sst.ann} has annual averages; \code{sst.mu} is the long-term temporal mean of the annual averages (so 1 value per grid cell).
}
\keyword{datasets}

