% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatValue.R
\name{is.species}
\alias{is.species}
\title{Is Species}
\usage{
is.species(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector of same length as x
}
\description{
Are there are least two words in this string?
}

