% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatStrat.R
\name{ll2km}
\alias{ll2km}
\title{Lon Lat to km}
\usage{
ll2km(x, y)
}
\arguments{
\item{x}{longitude}

\item{y}{latitude}
}
\value{
A named list with elements of Y (lat in km) and X (lon in km).
}
\description{
Convert longitude and latitude in degrees to the same in kilometers
}
\details{
Simple function to turn lon-lat distances (in degrees) into the kilometer equivalent. Given kilometers, easy to get areas, etc.
}

