% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatStrat.R
\name{ll2strat}
\alias{ll2strat}
\title{Lon lat to strat}
\usage{
ll2strat(lon, lat, gridSize = 1)
}
\arguments{
\item{lon}{numeric vector of longitude}

\item{lat}{numeric vector of latitude}

\item{gridSize}{numeric indicating fraction for rounding}
}
\value{
Character vector indicating the center of the defined stratum as the gridded lon and lat.
}
\description{
Convert longitude and latitude vectors to strata
}
\details{
\code{gridSize} is passed to \code{frac} in \link{\code{roundGrid}}. Defines a stratum based on rounding to the center of the nearest grid cell, of the specific size (\code{gridSize}).
}

