% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperMisc.R
\name{lu}
\alias{lu}
\title{Length Unique}
\usage{
lu(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{arguments to be passed to \code{\link{una}}}
}
\value{
an integer of the number of unique elements in \code{x}
}
\description{
The length of the vector of unique elements
}
\details{
Very convenient shorthand.
}
\examples{
lu(letters)
}

