% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatStrat.R
\name{makeStrat}
\alias{makeStrat}
\title{Make Strata}
\usage{
makeStrat(x, regName, doLots = NULL)
}
\arguments{
\item{x}{a data.table of trawl data}

\item{regName}{the name of the region}
}
\description{
Function to make strata for a region, examing missingness
}
\section{Warning}{

This function is not ready to be used. Saves figures, has hard-coded paths, looks for reference files outisde of package, etc.
}

