% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msomData.R
\name{msomData}
\alias{msomData}
\title{msomData}
\usage{
msomData(Data, n0 = 10, formula = year ~ stratum ~ K ~ spp,
  cov.vars = c(bt = "btemp", doy = "doy", yr = "year"), u.form = ~bt +
  I(bt^2), v.form = ~doy + I(doy^2) + yr, valueName = "abund",
  cov.by = c("year", "stratum", "K"))
}
\arguments{
\item{Data}{A data.table containing response variable and covariates}

\item{n0}{An integer indicating the number of never-observed species to add}

\item{formula}{A formula indicating the dimensions of the output array (only Default tested)}

\item{cov.vars}{A named character vector of covariates, where character elements correspond to columns \code{Data}, and names will be predictor names in model}

\item{u.form}{Formula for covariates to be included in U, the matrix of presence covariates}

\item{v.form}{Formula for covariates to be included in V, the matrix of detection covariates}

\item{valueName}{Character indicating column name containing response in \code{Data}}

\item{cov.by}{Chracter vector of dimension names for covariates; order matters, specify most specific last. Should be elements in \code{formula}, and columns in \code{Data}.}
}
\value{
A named list appropriate for use with msomStatic.stan
}
\description{
Format Data for MSOM
}
\details{
Only the defaults have been tested. Arguments that should be manipulated include \code{u.form}, \code{v.form}, \code{cov.vars}, and \code{n0}. Other values can be changed if they are only being adjusted to conform with column names in \code{Data}. Otherwise, it is unlikely that this function will behave as desired. In particular, the nesting of year, stratum, and K has to be the same (or whatever names are used).

Casting of \code{Data} is done via \code{trawlCast}. 0's are filled where appropriate, and NA's indicate no sampling. In the output, all NA's are replaced with 0's. However, the \code{nK} element of the output list indicates how many of the K are actually samples, and because K is not ordered, any 0's beyond the magnitude indicated in nK for that year-stratum combination are actually the NA's.
}

