% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.ai}
\alias{raw.ai}
\title{Raw Aleutian Islands}
\format{A data.table}
\usage{
raw.ai
}
\description{
Raw data set for the Aleutian Islands bottom trawl survey
}
\keyword{datasets}

