% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.ebs}
\alias{raw.ebs}
\title{Raw Eastern Berring Sea}
\format{A data.table}
\usage{
raw.ebs
}
\description{
Raw data set for the Eastern Berring Sea bottom trawl survey

\itemize{\item STRATUM the statistical stratum of the haul}

\itemize{\item LATITUDE latitude of the haul}

\itemize{\item LONGITUDE longitude of the haul, in western hemisphere degrees (for lon > 0, do lon-360)}

\itemize{\item STATION the station ID for the haul}

\itemize{\item YEAR year of haul}

\itemize{\item DATETIME the day and time of the haul}

\itemize{\item WTCPUE weight (mass) of the catch}

\itemize{\item NUMCPUE number of individuals caught per hectare in the haul}

\itemize{\item COMMON the common name of the organism sampled}

\itemize{\item SCIENTIFIC species scientific name; Genus species}

\itemize{\item SID species identification number}

\itemize{\item BOT_DEPTH the maximum depth of the water at the location of the haul}

\itemize{\item BOT_TEMP water temperature at the bottom at the location of the haul}

\itemize{\item SURF_TEMP water temperature at the surface at the location of the haul}

\itemize{\item VESSEL vessel ID}

\itemize{\item CRUISE cruise ID}

\itemize{\item HAUL the integer haul number within a cruise}

\itemize{\item Areakm2 the area of the statistical stratum (km2)}
}
\keyword{datasets}

