% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.gmex}
\alias{raw.gmex}
\title{Raw Gulf of Mexico}
\format{A data.table}
\usage{
raw.gmex
}
\description{
Raw data set for the Gulf of Mexico bottom trawl survey

\itemize{\item CRUISE_NO cruise ID}

\itemize{\item VESSEL vessel ID}

\itemize{\item BIO_BGS species identification number}

\itemize{\item P_STA_NO the station ID for the haul}

\itemize{\item CRUISE_NO cruise ID}

\itemize{\item P_STA_NO the station ID for the haul}

\itemize{\item BGSID record ID number}

\itemize{\item GENUS_BGS the genus of the species}

\itemize{\item SPEC_BGS the species name of the species}

\itemize{\item BGSCODE flags information about the biological sample}

\itemize{\item CNTEXP number of individuals in the whole net (may be extrapolated)}

\itemize{\item SELECT_BGS the weight (mass) of all items in the net (may be extrapolated)}

\itemize{\item GEAR_SIZE the dimension of the gear; for trawl net, the width of the mouth in ft}

\itemize{\item GEAR_TYPE code for the type of gear used}

\itemize{\item MESH_SIZE the size of the net mesh (inches of stretch)}

\itemize{\item MIN_FISH duration of the haul (how long the net was being towed)}

\itemize{\item TIME_ZN time zone}

\itemize{\item TIME_MIL starting time of the tow}

\itemize{\item E_LATD latitude of the haul; degree component}

\itemize{\item E_LATM latitude of the haul; minutes component}

\itemize{\item E_LOND longitude of the haul; degree component}

\itemize{\item E_LONM longitude of the haul; minutes component}

\itemize{\item DEPTH_SSTA depth of the haul}

\itemize{\item MO_DAY_YR date of the tow}

\itemize{\item E_LATD latitude of the haul; degree component}

\itemize{\item E_LATM latitude of the haul; minutes component}

\itemize{\item E_LOND longitude of the haul; degree component}

\itemize{\item E_LONM longitude of the haul; minutes component}

\itemize{\item TEMP_SSURF water temperature at the surface at the location of the haul}

\itemize{\item TEMP_BOT water temperature at the bottom at the location of the haul}

\itemize{\item VESSEL_SPD the speed of the vessel}

\itemize{\item TAXONOMIC species scientific name; Genus species}
}
\keyword{datasets}

