% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.goa}
\alias{raw.goa}
\title{Raw Gulf of Alaska}
\format{A data.table}
\usage{
raw.goa
}
\description{
Raw data set for the Gulf of Alaska bottom trawl survey
}
\keyword{datasets}

