% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.neus}
\alias{raw.neus}
\title{Raw Northeast US}
\format{A data.table}
\usage{
raw.neus
}
\description{
Raw data set for the Northeast US bottom trawl survey

\itemize{\item STRATUM the statistical stratum of the haul}

\itemize{\item SVSPP species identification number}

\itemize{\item YEAR year of haul}

\itemize{\item LAT latitude of the haul}

\itemize{\item LON longitude of the haul, in western hemisphere degrees (for lon > 0, do lon-360)}

\itemize{\item DEPTH the maximum depth of the water at the location of the haul}

\itemize{\item CRUISE6 cruise ID}

\itemize{\item STATION the station ID for the haul}

\itemize{\item BIOMASS the weight (mass) of all items in the net (may be extrapolated)}

\itemize{\item ABUNDANCE number of individuals in the whole net (may be extrapolated)}

\itemize{\item SCINAME species scientific name; Genus species}

\itemize{\item Areanmi2 the area of the statistical stratum (nmi2)}
}
\keyword{datasets}

