% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.sgulf}
\alias{raw.sgulf}
\title{Raw Southern Gulf of St. Lawrence}
\format{A data.table}
\usage{
raw.sgulf
}
\description{
Raw data set for the Southern Gulf of St. Lawrence bottom trawl survey
}
\keyword{datasets}

