% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.wcann}
\alias{raw.wcann}
\title{Raw West Coast Annual}
\format{A data.table}
\usage{
raw.wcann
}
\description{
Raw data set for the West Coast Annual bottom trawl survey

\itemize{\item Trawl.Id a unique identifier for the haul; vessel ID - cruise ID - haul number}

\itemize{\item Species species scientific name; Genus species}

\itemize{\item Haul.Weight..kg. the weight (mass) of all items in the net (may be extrapolated)}

\itemize{\item Survey.Cycle year of haul}

\itemize{\item Vessel vessel ID}

\itemize{\item Trawl.Date date of the tow}

\itemize{\item Trawl.Start.Time the day and time of the haul}

\itemize{\item Best.Latitude..dd. latitude of the haul}

\itemize{\item Best.Longitude..dd. longitude of the haul, in western hemisphere degrees (for lon > 0, do lon-360)}

\itemize{\item Best.Depth..m. the maximum depth of the water at the location of the haul}

\itemize{\item Trawl.Duration..min. duration of the haul (how long the net was being towed)}

\itemize{\item Area.Swept.by.the.Net..hectares. area sweap by the net (an index of effort)}

\itemize{\item Temperature.At.the.Gear..degs.C. water temperature at the bottom at the location of the haul}
}
\keyword{datasets}

