% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.wctri}
\alias{raw.wctri}
\title{Raw West Coast Triennial}
\format{A data.table}
\usage{
raw.wctri
}
\description{
Raw data set for the West Coast Triennial bottom trawl survey

\itemize{\item SPECIES_CODE species identification number}

\itemize{\item CRUISE cruise ID}

\itemize{\item HAUL_TYPE the integer haul number within a cruise}

\itemize{\item VESSEL vessel ID}

\itemize{\item CRUISE cruise ID}

\itemize{\item HAUL_TYPE the integer haul number within a cruise}

\itemize{\item WEIGHT the weight (mass) of all items in the net (may be extrapolated)}

\itemize{\item NUMBER_FISH number of individuals in the whole net (may be extrapolated)}

\itemize{\item HAUL_TYPE the integer haul number within a cruise}

\itemize{\item START_TIME starting time of the tow}

\itemize{\item DURATION duration of the haul (how long the net was being towed)}

\itemize{\item NET_WIDTH the dimension of the gear; for trawl net, the width of the mouth in ft}

\itemize{\item STRATUM the statistical stratum of the haul}

\itemize{\item STATIONID the station ID for the haul}

\itemize{\item BOTTOM_DEPTH the maximum depth of the water at the location of the haul}

\itemize{\item SURFACE_TEMPERATURE water temperature at the surface at the location of the haul}

\itemize{\item GEAR_TEMPERATURE water temperature at the bottom at the location of the haul}

\itemize{\item SPECIES_NAME species scientific name; Genus species}

\itemize{\item COMMON_NAME the common name of the organism sampled}
}
\keyword{datasets}

