% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.trawl.R
\name{read.trawl}
\alias{read.trawl}
\title{Read Trawl}
\usage{
read.trawl(reg = c("ai", "ebs", "gmex", "goa", "neus", "newf", "ngulf", "sa",
  "sgulf", "shelf", "wcann", "wctri"), zippath = file.path(system.file(package
  = "trawlData"), "extdata"), ...)
}
\arguments{
\item{reg}{the region name;  should be one of the following: \code{c("ai","ebs","gmex","goa", "neus", "newf", "sa", "sgulf", "shelf", "wcann", "wctri")}}

\item{zippath}{path to the zip file (or, for neus, to the folder containing a folder named 'neus')}

\item{...}{nothing}
}
\description{
Read in the raw trawl data sets for each region
}
\details{
These functions read in raw data sets that, with the exception of \code{neus}, are in .zip files. An easy way to use these functions on the raw files included with the package is to do \code{setwd(system.file(package="trawlData"))}, then run the script as-is. However, you are equally able to run the function as you would any other read function (like \code{read.csv}, e.g.). Just remember that the function operates on a folder or zip file, not on a single data file.
There is some use of regular expression to find appropriate file names. However, it is best to be sure that the files are named correctly (data files that come with this package are correctly named).
}
\examples{
\dontrun{
ai.data <- read.trawl(reg="ai", zippath=file.path(system.file(package="trawlData"),"inst/extdata"))
}
}

