% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperFile.R
\name{read.zip}
\alias{read.zip}
\title{Read Zip}
\usage{
read.zip(zipfile, pattern = "\\\\.csv$", SIMPLIFY = TRUE, use.fwf = FALSE,
  ...)
}
\arguments{
\item{zipfile}{name and path of a zip file containing files to read in}

\item{pattern}{regular expression indicating the file extensions to be read. Default is .csv files}

\item{SIMPLIFY}{logical, compress the contents into 2 data.table on output? To be safe, set FALSE}

\item{use.fwf}{logical, indicating whether or not to read in as fixed width file; is overridden if file extensions are .fwf}

\item{...}{other arguments to be passed on; in particular, the \code{cols} argument of column width if a fixed width file}
}
\value{
a data.table, or list of data.tables. The name of each element of the list is the name of the file within the .zip file.
}
\description{
Read in all files contained in a zip file
}
\details{
The function works by unzipping a file into a temporary directory, reading in all the files in that temporarily unzipped folder, and then deleting the temporary folder and all of its contents. Thus, is a handy way of keep all of your files stored as compressed files, but being able to access them. Obviously comes at the performance cost of having to do all of the unzipping and deleting.
Uses data.table's \code{\link{fread}} in most cases, except for .fwf formats, in which case \code{\link{fread.fwf}} is used.
}

