% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatValue.R
\name{rm9s}
\alias{rm9s}
\title{Remove 9's}
\usage{
rm9s(x)
}
\arguments{
\item{x}{A data.table}
}
\value{
Nothing, but affects data.table passed as \code{x}.
}
\description{
Remove 9's and switch them to NA
}
\details{
All instances of -9999 (numeric or integer) are replaced as NA's of the appropriate class. Checks also for class "integer64".
}

