% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperMisc.R
\name{s2c}
\alias{s2c}
\title{String to call}
\usage{
s2c(x, type = "list")
}
\arguments{
\item{x}{a character string (e.g., of column names in a data.table)}

\item{type}{the class resulting from the evaluation of the call returned by s2c}
}
\value{
a list
}
\description{
Convert a character vector to a call of the desired class
}
\details{
e.g., convert \code{c("column1","column2")} to \code{list(column1, column2)}
intended for use in a data.table (see \code{\link{data.table}})
}
\examples{
library(data.table)
first <- data.table(cray=sample(letters,4),"one"=c(1,2,3,4), "two"=c(1,3,5,7))
second <- data.table("cray"=sample(letters,35, TRUE))
oc <- CJ(one=first[,(one)], cray=second[[1]])
sub <- s2c(c("one","cray"))
oc[,eval(sub)]
}

