% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spp.corr1}
\alias{spp.corr1}
\title{Corrected Species Names (legacy)
Species names that had been corrected in older version of code}
\format{A dim = 3990 x 2 data.table data.frame:
\tabular{rlll}{
[,1] \tab spp \tab character \tab insert_description_here\cr
[,2] \tab sppCorr \tab character \tab insert_description_here\cr
}}
\usage{
spp.corr1
}
\description{
Corrected Species Names (legacy)
Species names that had been corrected in older version of code
}
\details{
Used to construct spp.key initially, but should no longer be needed. Provided for reproducibility of older results regarding taxonomy.
Now superceded by \code{\link{getSppData}}. It is recommended that \code{\link{spp.key}} be used for taxonomic reference.
}
\seealso{
\code{\link{getCmmn}} \code{\link{getSpp}} \code{\link{getTax}} \code{\link{spp.key}}
}
\keyword{datasets}

