% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sppImg.R
\name{sppImg}
\alias{sppImg}
\title{Species Image}
\usage{
sppImg(x, common = NULL, ...)
}
\arguments{
\item{x}{the name of the species (Genus species)}

\item{common}{The species' common name; if NULL, no common name printed}

\item{...}{Arguments to be passed to \code{\link{image}}}
}
\value{
Numeric vector of length two indicating the number of pixes horizontally and vertically; or if no match, returns NULL and prints a message.
}
\description{
Plot an image of a species from the package database of images
}
\examples{
pick <- function(x, n){
	(x)\%in\%sample(unique(x),n)
}

library(rgdal)
dev.new(width=8, height=7)
par(mar=c(0.5,0.5,1,0.5), oma=c(0,0,0,0), ps=8, cex=1, mfrow=c(5,5))
clean.ebs[Picture=="y"][pick(spp,25),sppImg(unique(spp),unique(common),side=3, adj=0.1, line=-0.85, xpd=TRUE),by="spp"]
}

