% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperStat.R
\name{sumna}
\alias{sumna}
\title{Sum NA}
\usage{
sumna(x)
}
\arguments{
\item{x}{numbers to be summed}
}
\description{
Sum where NA's are removed, but if all are NA, return NA (not 0)
}
\seealso{
\code{\link{sumna}}, \code{\link{meanna}}, \code{\link{wtAvg}}, \code{\link{fill.mean}}
}

