% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawlTrim.R
\name{trawlTrim}
\alias{trawlTrim}
\title{Trim Trawl Data}
\usage{
trawlTrim(reg, ...)
}
\arguments{
\item{reg}{a list of data.tables or a character vector of regions to be in data set. If a list of data.tables, trimming and combining is performed on those data.tables. If a charcter vector of region names, package data sets by the name of \code{paste0("clean.", reg)} are used. If nothing is supplied, \code{reg} is a character vector of all region names.}

\item{...}{arguments to be passed to \code{\link{clean.trimCol}}}
}
\description{
Get a trimmed data set for one or more regions
}
\details{
This function drops columns and deletes rows. Column dropping is performed by \code{\link{clean.trimCol}}, rows to drop determined by flagging performed by \code{\link{clean.trimRow}}.  

If multiple regions are supplied to \code{reg}, those regions will be collated via \code{\link{rbindlist}}. This is possible because trimming columns standardizes content. Additionally, although \code{\link{clean.trimRow}} adds the \code{keep.row} column, this function will actually drop those rows for which \code{keep.row} is FALSE.
}
\examples{
# default usage
trawlTrim(c("ebs","goa"))
# usage to get an extra column
trawlTrim(c("ebs","shelf"), c.add="WIND")

}

