% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperMisc.R
\name{una}
\alias{una}
\title{Unique non-NA}
\usage{
una(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{a vector from which to take unique values}

\item{na.rm}{Default FALSE; remove NA's?}

\item{...}{other arguments to be passed to \code{\link{unique}}}
}
\description{
Unique elements, excluding NA's
}
\examples{
una(NA)
una(NA, na.rm=TRUE)
una(1:3)
una(c(1:3,NA))
una(c(1:3,NA), na.rm=TRUE)

}
\seealso{
\code{\link{lu}}
}

