% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperStat.R
\name{wtAvg}
\alias{wtAvg}
\title{Weighted Average, NA}
\usage{
wtAvg(x, y)
}
\arguments{
\item{x}{the value to be average over}

\item{y}{the value to be transformed into weights}
}
\description{
Take the weighted average, with NA's removed, but if all are NA, return NA (not 0)
}
\seealso{
\code{\link{sumna}}, \code{\link{meanna}}, \code{\link{wtAvg}}, \code{\link{fill.mean}}
}

