% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.ai}
\alias{clean.ai}
\title{Clean Aleutian Islands}
\format{A dim = 118661 x 48 data.table data.frame:
\tabular{rlll}{
[,1] \tab ref \tab character \tab reference taxonomic ID from raw data\cr
[,2] \tab stratum \tab character \tab the statistical stratum of the haul\cr
[,3] \tab lat \tab numeric \tab latitude of the haul\cr
[,4] \tab lon \tab numeric \tab longitude of the haul, in western hemisphere degrees (for lon > 0, do lon-360)\cr
[,5] \tab station \tab character \tab the station ID for the haul\cr
[,6] \tab year \tab integer \tab year of haul\cr
[,7] \tab datetime \tab c("POSIXct", "POSIXt") \tab the day and time of the haul\cr
[,8] \tab wtcpue \tab numeric \tab weight (mass) of the catch\cr
[,9] \tab cntcpue \tab numeric \tab number of individuals caught per hectare in the haul\cr
[,10] \tab SID \tab integer \tab species identification number\cr
[,11] \tab depth \tab numeric \tab the maximum depth of the water at the location of the haul\cr
[,12] \tab btemp \tab numeric \tab water temperature at the bottom at the location of the haul\cr
[,13] \tab stemp \tab numeric \tab water temperature at the surface at the location of the haul\cr
[,14] \tab vessel \tab character \tab vessel ID\cr
[,15] \tab cruise \tab character \tab cruise ID\cr
[,16] \tab haul \tab character \tab the integer haul number within a cruise\cr
[,17] \tab stratumarea \tab numeric \tab the area of the statistical stratum (km2)\cr
[,18] \tab haulid \tab character \tab a unique identifier for the haul; vessel ID - cruise ID - haul number\cr
[,19] \tab season \tab character \tab insert_description_here\cr
[,20] \tab reg \tab character \tab survey region\cr
[,21] \tab val.src \tab character \tab indicates the degree of 'fuzziness' required to find a match to ref in a data.base of taxonomic information; m1 indicates perfect match, m2 indicates that capitalization, whitespace, etc (see \code{\link{cull}}) needed to be adjusted, m3 indicates that \code{\link{agrep}} was used, and m4 means that measures in both m2 and m3 were taken to find the match. See \code{\link{match.tbl}}.\cr
[,22] \tab tbl.row \tab integer \tab the row in the taxonomic data base where a match was found. See \code{\link{match.tbl}}.\cr
[,23] \tab mtch.src \tab integer \tab the database containing the match; 1 is taxInfo, 2 is spp.corr1, 3 is getSppData; if matches are found in multiple sources, a match to 1 takes precedence over 1 & 2, and 2 over 3. See \code{\link{match.tbl}}.\cr
[,24] \tab tax.src \tab character \tab informs source of taxonomic correction of ref to spp and other tax info; is taxInfo if found from manually checked spreadsheet\cr
[,25] \tab spp \tab character \tab species scientific name; Genus species\cr
[,26] \tab common \tab character \tab the common name of the organism sampled\cr
[,27] \tab taxLvl \tab character \tab the most specific level of classification indicated by spp\cr
[,28] \tab species \tab character \tab the species name of the species\cr
[,29] \tab genus \tab character \tab the genus of the species\cr
[,30] \tab family \tab character \tab taxonomic family\cr
[,31] \tab order \tab character \tab taxonomic order\cr
[,32] \tab class \tab character \tab taxonomic class\cr
[,33] \tab superclass \tab character \tab taxonomic superclass\cr
[,34] \tab subphylum \tab character \tab taxonomic subphylum\cr
[,35] \tab phylum \tab character \tab insert_description_here\cr
[,36] \tab kingdom \tab character \tab taxonomic kingdom\cr
[,37] \tab trophicDiet \tab character \tab source of trophic level from Fish Base or Sea Life Base; 'y' means it was from this source\cr
[,38] \tab trophicOrig \tab character \tab from Fish Base or Sea Life Base; was the trophic level estimated from an 'Original sample'?\cr
[,39] \tab Picture \tab character \tab Is there a picture of this critter assoicated with the package? Note: this isn't always accurate\cr
[,40] \tab trophicLevel \tab numeric \tab the trophic level from Fish Base or Sea Life Base\cr
[,41] \tab trophicLevel.se \tab numeric \tab the standard error of the trophic level\cr
[,42] \tab tax.src2 \tab character \tab informs source of taxonomic correct; the name of a source of taxonomic information other than taxInfo (other than manual entries)\cr
[,43] \tab conflict \tab logical \tab for a given 'spp' value in spp.key, was there a conflict in the other taxonomic columns? E.g., a single spp corresponding to multiple common names; also TRUE if different ref values were found in different databases (affecting the val.src, tbl.row, mtch.src, tax.src, etc columns), but then the refs converged to same spp -- that would not necessarily be an error, but might deserve checking\cr
[,44] \tab flag \tab character \tab flag related to correcting taxonomic information; relates to automated input, potential errors, and signature of people or methods that have made corrections to spp.key\cr
[,45] \tab website \tab character \tab URL reference for taxonomic information\cr
[,46] \tab website2 \tab character \tab secondary URL reference for taxonomic information; often used when website finds the name for a spp name as entered, but website2 points to the most up-to-date scientific name\cr
[,47] \tab keep.row \tab logical \tab Column indicating whether or not the row show likely be excluded\cr
[,48] \tab row_flag \tab character \tab if keep.row is TRUE, why was this row flagged to be dropped?\cr
}}
\usage{
clean.ai
}
\description{
Clean data from the Aleutian Islands bottom trawl
}
\keyword{datasets}

