% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.ebs}
\alias{clean.ebs}
\title{Clean Eastern Berring Sea}
\format{A data.table}
\usage{
clean.ebs
}
\description{
Clean data set for the Eastern Berring Sea bottom trawl survey
}
\keyword{datasets}

