% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.gmex}
\alias{clean.gmex}
\title{Clean Gulf of Mexico}
\format{A data.table}
\usage{
clean.gmex
}
\description{
Clean data set for the Gulf of Mexico bottom trawl survey
}
\keyword{datasets}

