% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.goa}
\alias{clean.goa}
\title{Clean Gulf of Alaska}
\format{A data.table}
\usage{
clean.goa
}
\description{
Clean data set for the Gulf of Alaska bottom trawl survey
}
\keyword{datasets}

