% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.goa}
\alias{clean.goa}
\title{Clean Gulf of Alaska}
\format{A dim = 197067 x 41 data.table data.frame:
\tabular{rlll}{
[,1] \tab ref \tab character \tab insert_description_here\cr
[,2] \tab stratum \tab character \tab insert_description_here\cr
[,3] \tab lat \tab numeric \tab insert_description_here\cr
[,4] \tab lon \tab numeric \tab insert_description_here\cr
[,5] \tab station \tab character \tab insert_description_here\cr
[,6] \tab year \tab character \tab insert_description_here\cr
[,7] \tab datetime \tab c("POSIXct", "POSIXt") \tab insert_description_here\cr
[,8] \tab wtcpue \tab numeric \tab insert_description_here\cr
[,9] \tab cntcpue \tab numeric \tab insert_description_here\cr
[,10] \tab SID \tab integer \tab insert_description_here\cr
[,11] \tab depth \tab numeric \tab insert_description_here\cr
[,12] \tab btemp \tab numeric \tab insert_description_here\cr
[,13] \tab stemp \tab numeric \tab insert_description_here\cr
[,14] \tab vessel \tab character \tab insert_description_here\cr
[,15] \tab cruise \tab character \tab insert_description_here\cr
[,16] \tab haul \tab character \tab insert_description_here\cr
[,17] \tab stratumarea \tab numeric \tab insert_description_here\cr
[,18] \tab haulid \tab character \tab insert_description_here\cr
[,19] \tab season \tab character \tab insert_description_here\cr
[,20] \tab reg \tab character \tab insert_description_here\cr
[,21] \tab val.src \tab character \tab insert_description_here\cr
[,22] \tab tbl.row \tab numeric \tab insert_description_here\cr
[,23] \tab mtch.src \tab numeric \tab insert_description_here\cr
[,24] \tab spp \tab character \tab insert_description_here\cr
[,25] \tab common \tab character \tab insert_description_here\cr
[,26] \tab taxLvl \tab character \tab insert_description_here\cr
[,27] \tab species \tab character \tab insert_description_here\cr
[,28] \tab genus \tab character \tab insert_description_here\cr
[,29] \tab family \tab character \tab insert_description_here\cr
[,30] \tab order \tab character \tab insert_description_here\cr
[,31] \tab class \tab character \tab insert_description_here\cr
[,32] \tab superclass \tab character \tab insert_description_here\cr
[,33] \tab subphylum \tab character \tab insert_description_here\cr
[,34] \tab phylum \tab character \tab insert_description_here\cr
[,35] \tab kingdom \tab character \tab insert_description_here\cr
[,36] \tab trophicDiet \tab character \tab insert_description_here\cr
[,37] \tab trophicOrig \tab character \tab insert_description_here\cr
[,38] \tab Picture \tab character \tab insert_description_here\cr
[,39] \tab trophicLevel \tab numeric \tab insert_description_here\cr
[,40] \tab trophicLevel.se \tab numeric \tab insert_description_here\cr
[,41] \tab keep.row \tab logical \tab insert_description_here\cr
}}
\usage{
clean.goa
}
\description{
Clean data set for the Gulf of Alaska bottom trawl survey
}
\keyword{datasets}

