% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.neus}
\alias{clean.neus}
\title{Clean Northeast US}
\format{A data.table}
\usage{
clean.neus
}
\description{
Clean data set for the Northeast US bottom trawl survey
}
\keyword{datasets}

