% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.newf}
\alias{clean.newf}
\title{Clean Newfoundland}
\format{A data.table}
\usage{
clean.newf
}
\description{
Clean data set for the Newfoundland bottom trawl survey
}
\keyword{datasets}

