% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.sgulf}
\alias{clean.sgulf}
\title{Clean Southern Gulf of St. Lawrence}
\format{A data.table}
\usage{
clean.sgulf
}
\description{
Clean data set for the Southern Gulf of St. Lawrence bottom trawl survey
}
\keyword{datasets}

