% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.shelf}
\alias{clean.shelf}
\title{Clean Scotian Shelf}
\format{A data.table}
\usage{
clean.shelf
}
\description{
Clean data set for the Scotian Shelf bottom trawl survey
}
\keyword{datasets}

