% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.wcann}
\alias{clean.wcann}
\title{Clean West Coast Annual}
\format{A data.table}
\usage{
clean.wcann
}
\description{
Clean data set for the West Coast Annual bottom trawl survey
}
\keyword{datasets}

