% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clean.wctri}
\alias{clean.wctri}
\title{Clean West Coast Triennial}
\format{A dim = 108138 x 79 data.table data.frame:
\tabular{rlll}{
[,1] \tab ref \tab character \tab reference taxonomic ID from raw data\cr
[,2] \tab SID \tab integer \tab species identification number\cr
[,3] \tab CRUISEJOIN \tab integer \tab insert_description_here\cr
[,4] \tab HAULJOIN \tab integer \tab insert_description_here\cr
[,5] \tab REGION \tab character \tab insert_description_here\cr
[,6] \tab vessel \tab character \tab vessel ID\cr
[,7] \tab cruise \tab character \tab cruise ID\cr
[,8] \tab haul \tab character \tab the integer haul number within a cruise\cr
[,9] \tab CATCHJOIN \tab integer \tab insert_description_here\cr
[,10] \tab weight \tab numeric \tab the weight (mass) of all items in the net (may be extrapolated)\cr
[,11] \tab cnt \tab numeric \tab number of individuals in the whole net (may be extrapolated)\cr
[,12] \tab SUBSAMPLE_CODE \tab logical \tab insert_description_here\cr
[,13] \tab VOUCHER \tab integer \tab insert_description_here\cr
[,14] \tab AUDITJOIN \tab integer \tab insert_description_here\cr
[,15] \tab HAUL_TYPE \tab integer \tab insert_description_here\cr
[,16] \tab PERFORMANCE \tab numeric \tab insert_description_here\cr
[,17] \tab datetime \tab c("POSIXct", "POSIXt") \tab the day and time of the haul\cr
[,18] \tab towduration \tab numeric \tab the duration (time) for the tow\cr
[,19] \tab towdistance \tab numeric \tab insert_description_here\cr
[,20] \tab gearsize \tab numeric \tab the dimension of the gear; for trawl net, the width of the mouth in ft\cr
[,21] \tab NET_MEASURED \tab character \tab insert_description_here\cr
[,22] \tab NET_HEIGHT \tab numeric \tab insert_description_here\cr
[,23] \tab stratum \tab character \tab the statistical stratum of the haul\cr
[,24] \tab lat.start \tab numeric \tab insert_description_here\cr
[,25] \tab lat.end \tab numeric \tab insert_description_here\cr
[,26] \tab lon.start \tab numeric \tab insert_description_here\cr
[,27] \tab lon.end \tab numeric \tab insert_description_here\cr
[,28] \tab station \tab character \tab the station ID for the haul\cr
[,29] \tab GEAR_DEPTH \tab integer \tab insert_description_here\cr
[,30] \tab depth \tab numeric \tab the maximum depth of the water at the location of the haul\cr
[,31] \tab BOTTOM_TYPE \tab character \tab insert_description_here\cr
[,32] \tab stemp \tab numeric \tab water temperature at the surface at the location of the haul\cr
[,33] \tab btemp \tab numeric \tab water temperature at the bottom at the location of the haul\cr
[,34] \tab WIRE_LENGTH \tab integer \tab insert_description_here\cr
[,35] \tab geartype \tab character \tab code for the type of gear used\cr
[,36] \tab ACCESSORIES \tab character \tab insert_description_here\cr
[,37] \tab SUBSAMPLE \tab character \tab insert_description_here\cr
[,38] \tab REVISION \tab character \tab insert_description_here\cr
[,39] \tab BS \tab character \tab insert_description_here\cr
[,40] \tab GOA \tab character \tab insert_description_here\cr
[,41] \tab WC \tab character \tab insert_description_here\cr
[,42] \tab haulid \tab character \tab a unique identifier for the haul; vessel ID - cruise ID - haul number\cr
[,43] \tab year \tab numeric \tab year of haul\cr
[,44] \tab season \tab character \tab insert_description_here\cr
[,45] \tab lat \tab numeric \tab latitude of the haul\cr
[,46] \tab lon \tab numeric \tab longitude of the haul, in western hemisphere degrees (for lon > 0, do lon-360)\cr
[,47] \tab stratumarea \tab numeric \tab the area of the statistical stratum (km2)\cr
[,48] \tab towarea \tab numeric \tab insert_description_here\cr
[,49] \tab effort \tab numeric \tab some measure of the effort for this catch; could be area, time, speed, net size, or some combination. When possible, was converted to area swept by net during the entire tow\cr
[,50] \tab wtcpue \tab numeric \tab weight (mass) of the catch\cr
[,51] \tab cntcpue \tab numeric \tab number of individuals caught per hectare in the haul\cr
[,52] \tab reg \tab character \tab survey region\cr
[,53] \tab val.src \tab character \tab indicates the degree of 'fuzziness' required to find a match to ref in a data.base of taxonomic information; m1 indicates perfect match, m2 indicates that capitalization, whitespace, etc (see \code{\link{cull}}) needed to be adjusted, m3 indicates that \code{\link{agrep}} was used, and m4 means that measures in both m2 and m3 were taken to find the match\cr
[,54] \tab tbl.row \tab integer \tab the row in the taxonomic data base where a match was found\cr
[,55] \tab mtch.src \tab integer \tab the database containing the match; 1 is taxInfo, 2 is spp.corr1, 3 is getSppData; if matches are found in multiple sources, a match to 1 takes precedence over 1 & 2, and 2 over 3\cr
[,56] \tab tax.src \tab character \tab informs source of taxonomic correction of ref to spp and other tax info; is taxInfo if found from manually checked spreadsheet\cr
[,57] \tab spp \tab character \tab species scientific name; Genus species\cr
[,58] \tab common \tab character \tab the common name of the organism sampled\cr
[,59] \tab taxLvl \tab character \tab the most specific level of classification indicated by spp\cr
[,60] \tab species \tab character \tab the species name of the species\cr
[,61] \tab genus \tab character \tab the genus of the species\cr
[,62] \tab family \tab character \tab taxonomic family\cr
[,63] \tab order \tab character \tab taxonomic order\cr
[,64] \tab class \tab character \tab taxonomic class\cr
[,65] \tab superclass \tab character \tab taxonomic superclass\cr
[,66] \tab subphylum \tab character \tab taxonomic subphylum\cr
[,67] \tab phylum \tab character \tab insert_description_here\cr
[,68] \tab kingdom \tab character \tab taxonomic kingdom\cr
[,69] \tab trophicDiet \tab character \tab source of trophic level from Fish Base or Sea Life Base; 'y' means it was from this source\cr
[,70] \tab trophicOrig \tab character \tab from Fish Base or Sea Life Base; was the trophic level estimated from an 'Original sample'?\cr
[,71] \tab Picture \tab character \tab Is there a picture of this critter assoicated with the package? Note: this isn't always accurate\cr
[,72] \tab trophicLevel \tab numeric \tab the trophic level from Fish Base or Sea Life Base\cr
[,73] \tab trophicLevel.se \tab numeric \tab the standard error of the trophic level\cr
[,74] \tab tax.src2 \tab character \tab informs source of taxonomic correct; the name of a source of taxonomic information other than taxInfo (other than manual entries)\cr
[,75] \tab conflict \tab logical \tab for a given 'spp' value in spp.key, was there a conflict in the other taxonomic columns? E.g., a single spp corresponding to multiple common names; also TRUE if different ref values were found in different databases (affecting the val.src, tbl.row, mtch.src, tax.src, etc columns), but then the refs converged to same spp -- that would not necessarily be an error, but might deserve checking\cr
[,76] \tab flag \tab character \tab flag related to correcting taxonomic information; relates to automated input, potential errors, and signature of people or methods that have made corrections to spp.key\cr
[,77] \tab website \tab character \tab URL reference for taxonomic information\cr
[,78] \tab website2 \tab character \tab secondary URL reference for taxonomic information; often used when website finds the name for a spp name as entered, but website2 points to the most up-to-date scientific name\cr
[,79] \tab keep.row \tab logical \tab Column indicating whether or not the row show likely be excluded\cr
}}
\usage{
clean.wctri
}
\description{
Clean data set for the West Coast Triennial bottom trawl survey
}
\keyword{datasets}

