% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tax.getCmmn.R
\name{getCmmn}
\alias{getCmmn}
\title{Get Common}
\usage{
getCmmn(u.sppCorr)
}
\arguments{
\item{u.sppCorr}{character vector of Latin names to lookup

Uses rfishbase (this is new from original implementation), and only uses taxize if rfishbase fails to find a match (but taxize is very slow sometimes). Furture versions should use the sealifebase implementation of rfishbase, which I haven't figured out/ tested yet.}
}
\value{
a data.table with two columns: the species name searched, and the common name
}
\description{
Get species' common names
}
\examples{
getCmmn(c("Paralichthys dentatus", "Pomatomus saltatrix","Gadus morhua","Chionoecetes tanneri"))
}

