% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{getCmmnData}
\alias{getCmmnData}
\title{Common Names}
\format{A dim = 3548 x 2 data.table data.frame:
\tabular{rlll}{
[,1] \tab sppCorr \tab character \tab insert_description_here\cr
[,2] \tab common \tab character \tab the common name of the organism sampled\cr
}}
\usage{
getCmmnData
}
\description{
Species's common names matched to scientific names
}
\keyword{datasets}

