% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{getCmmnData}
\alias{getCmmnData}
\title{Common Names  
  
Species's common names matched to scientific names}
\format{A dim = 3548 x 2 data.table data.frame:  
\tabular{rlll}{
[,1] \tab sppCorr \tab character \tab insert_description_here\cr
[,2] \tab common \tab character \tab the common name of the organism sampled\cr
}}
\usage{
getCmmnData
}
\description{
Common Names  
  
Species's common names matched to scientific names
}
\keyword{datasets}

