% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{getCmmnData}
\alias{getCmmnData}
\title{Common Names
Species' common names matched to scientific names}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	3548 obs. of  2 variables:
 $ sppCorr: chr  "Abietinaria" "Abietinaria greenei" "Abisa" "Ablennes hians" ...
 $ common : chr  NA NA NA "agujón sable" ...
 - attr(*, ".internal.selfref")=<externalptr> 
 - attr(*, "sorted")= chr "sppCorr"
}}
\usage{
getCmmnData
}
\description{
Common Names
Species' common names matched to scientific names
}
\seealso{
\code{\link{getCmmn}}
}
\keyword{datasets}

