% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{getTaxData}
\alias{getTaxData}
\title{Taxonomic Classification
Taxonomic classification matched to scientific names}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	3548 obs. of  11 variables:
 $ sppCorr   : chr  "Abietinaria" "Abietinaria greenei" "Abisa" "Ablennes hians" ...
 $ taxLvl    : chr  "genus" NA NA "species" ...
 $ species   : chr  NA NA NA "Ablennes hians" ...
 $ genus     : chr  "Abietinaria" NA NA "Ablennes" ...
 $ family    : chr  "Sertulariidae" NA NA "Belonidae" ...
 $ order     : chr  "Leptothecata" NA NA "Beloniformes" ...
 $ class     : chr  "Hydrozoa" NA NA "Actinopteri" ...
 $ superclass: chr  NA NA NA "Actinopterygii" ...
 $ subphylum : chr  NA NA NA "Craniata" ...
 $ phylum    : chr  "Cnidaria" NA NA "Chordata" ...
 $ kingdom   : chr  "Metazoa" NA NA "Metazoa" ...
 - attr(*, ".internal.selfref")=<externalptr> 
 - attr(*, "sorted")= chr "sppCorr"
}}
\usage{
getTaxData
}
\description{
Taxonomic Classification
Taxonomic classification matched to scientific names
}
\seealso{
\code{\link{getTax}}
}
\keyword{datasets}

