% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helperStat.R
\name{meanna}
\alias{meanna}
\title{Mean NA}
\usage{
meanna(x)
}
\arguments{
\item{x}{take the mean of these numbers}
}
\description{
Mean where NA's are removed, but if all are NA, return NA (not 0)
}
\seealso{
\code{\link{sumna}}, \code{\link{meanna}}, \code{\link{wtAvg}}, \code{\link{fill.mean}}
}

