% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.ai}
\alias{raw.ai}
\title{Raw Aleutian Islands  
  
Raw data from the Aleutian Islands bottom trawl}
\format{A dim = 118661 x 18 data.table data.frame:  
\tabular{rlll}{
[,1] \tab STRATUM \tab integer \tab cleans up to \code{stratum}, the statistical stratum of the haul\cr
[,2] \tab LATITUDE \tab numeric \tab cleans up to \code{lat}, latitude of the haul\cr
[,3] \tab LONGITUDE \tab numeric \tab cleans up to \code{lon}, longitude of the haul, in western hemisphere degrees (for lon > 0, do lon-360)\cr
[,4] \tab STATION \tab character \tab cleans up to \code{station}, the station ID for the haul\cr
[,5] \tab YEAR \tab integer \tab cleans up to \code{year}, year of haul\cr
[,6] \tab DATETIME \tab character \tab cleans up to \code{datetime}, the day and time of the haul\cr
[,7] \tab WTCPUE \tab numeric \tab cleans up to \code{wtcpue}, weight (mass) of the catch\cr
[,8] \tab NUMCPUE \tab numeric \tab cleans up to \code{cntcpue}, number of individuals caught per hectare in the haul\cr
[,9] \tab COMMON \tab character \tab cleans up to \code{common}, the common name of the organism sampled\cr
[,10] \tab SCIENTIFIC \tab character \tab cleans up to \code{spp}, species scientific name; Genus species\cr
[,11] \tab SID \tab integer \tab species identification number\cr
[,12] \tab BOT_DEPTH \tab integer \tab cleans up to \code{depth}, the maximum depth of the water at the location of the haul\cr
[,13] \tab BOT_TEMP \tab numeric \tab cleans up to \code{btemp}, water temperature at the bottom at the location of the haul\cr
[,14] \tab SURF_TEMP \tab numeric \tab cleans up to \code{stemp}, water temperature at the surface at the location of the haul\cr
[,15] \tab VESSEL \tab integer \tab cleans up to \code{vessel}, vessel ID\cr
[,16] \tab CRUISE \tab integer \tab cleans up to \code{cruise}, cruise ID\cr
[,17] \tab HAUL        \tab character \tab cleans up to \code{haul}, the integer haul number within a cruise\cr
[,18] \tab Areakm2 \tab integer \tab cleans up to \code{stratumarea}, the area of the statistical stratum (km2)\cr
}}
\usage{
raw.ai
}
\description{
Raw Aleutian Islands  
  
Raw data from the Aleutian Islands bottom trawl
}
\keyword{datasets}

