% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.ebs}
\alias{raw.ebs}
\title{\itemize{\item STRATUM the statistical stratum of the haul}}
\format{A data.table}
\usage{
raw.ebs
}
\description{
\itemize{\item STRATUM the statistical stratum of the haul}
}
\keyword{datasets}

