% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.ebs}
\alias{raw.ebs}
\title{Raw Eastern Berring Sea}
\format{A data.table}
\usage{
raw.ebs
}
\description{
Raw data set for the Eastern Berring Sea bottom trawl survey
}
\keyword{datasets}

