% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.gmex}
\alias{raw.gmex}
\title{Raw Gulf of Mexico}
\format{A data.table}
\usage{
raw.gmex
}
\description{
Raw data set for the Gulf of Mexico bottom trawl survey
}
\keyword{datasets}

