% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.neus}
\alias{raw.neus}
\title{Raw Northeast US}
\format{A data.table}
\usage{
raw.neus
}
\description{
Raw data set for the Northeast US bottom trawl survey
}
\keyword{datasets}

