% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.sa}
\alias{raw.sa}
\title{Raw South Atlantic (Southeast US)  
  
Raw data set for the South Atlantic (Southeast US) bottom trawl survey}
\format{A dim = 351427 x 62 data.table data.frame:  
\tabular{rlll}{
[,1] \tab COLLECTIONNUMBER \tab integer \tab insert_description_here\cr
[,2] \tab PROJECTNAME \tab character \tab insert_description_here\cr
[,3] \tab PROJECTAGENCY \tab character \tab insert_description_here\cr
[,4] \tab DATE \tab character \tab cleans up to \code{date}, date of the tow\cr
[,5] \tab EVENTNAME \tab integer \tab cleans up to \code{haulid}, a unique identifier for the haul; vessel ID - cruise ID - haul number\cr
[,6] \tab VESSELNAME \tab character \tab cleans up to \code{vessel}, vessel ID\cr
[,7] \tab GEARNAME \tab character \tab insert_description_here\cr
[,8] \tab GEARCODE \tab integer \tab insert_description_here\cr
[,9] \tab SPECIESCODE \tab integer64 \tab insert_description_here\cr
[,10] \tab MRRI_CODE \tab character \tab insert_description_here\cr
[,11] \tab SPECIESSCIENTIFICNAME \tab character \tab cleans up to \code{spp}, species scientific name; Genus species\cr
[,12] \tab SPECIESCOMMONNAME \tab character \tab cleans up to \code{common}, the common name of the organism sampled\cr
[,13] \tab NUMBERTOTAL \tab integer \tab cleans up to \code{cnt}, number of individuals in the whole net (may be extrapolated)\cr
[,14] \tab SPECIESTOTALWEIGHT \tab numeric \tab cleans up to \code{weight}, the weight (mass) of all items in the net (may be extrapolated)\cr
[,15] \tab SPECIESSUBWEIGHT \tab numeric \tab insert_description_here\cr
[,16] \tab SPECIESWGTPROCESSED \tab numeric \tab insert_description_here\cr
[,17] \tab WEIGHTMETHODDESC \tab character \tab insert_description_here\cr
[,18] \tab ORGWTUNITS \tab character \tab insert_description_here\cr
[,19] \tab EFFORT \tab numeric \tab cleans up to \code{effort}, some measure of the effort for this catch; could be area, time, speed, net size, or some combination. When possible, was converted to area swept by net during the entire tow\cr
[,20] \tab CATCHSUBSAMPLED \tab logical \tab insert_description_here\cr
[,21] \tab CATCHWEIGHT \tab numeric \tab insert_description_here\cr
[,22] \tab CATCHSUBWEIGHT \tab numeric \tab insert_description_here\cr
[,23] \tab TIMESTART \tab character \tab cleans up to \code{time}, starting time of the tow\cr
[,24] \tab DURATION \tab integer \tab cleans up to \code{towduration}, the duration (time) for the tow\cr
[,25] \tab TOWTYPETEXT \tab character \tab insert_description_here\cr
[,26] \tab LOCATION \tab character \tab insert_description_here\cr
[,27] \tab REGION \tab character \tab insert_description_here\cr
[,28] \tab DEPTHZONE \tab character \tab insert_description_here\cr
[,29] \tab ACCSPGRIDCODE \tab character \tab insert_description_here\cr
[,30] \tab STATIONCODE \tab character \tab cleans up to \code{stratum}, the statistical stratum of the haul\cr
[,31] \tab EVENTTYPEDESCRIPTION \tab character \tab insert_description_here\cr
[,32] \tab TEMPSURFACE \tab numeric \tab cleans up to \code{stemp}, water temperature at the surface at the location of the haul\cr
[,33] \tab TEMPBOTTOM \tab numeric \tab cleans up to \code{btemp}, water temperature at the bottom at the location of the haul\cr
[,34] \tab SALINITYSURFACE \tab numeric \tab cleans up to \code{ssalin}, surface salinity\cr
[,35] \tab SALINITYBOTTOM \tab numeric \tab cleans up to \code{bsalin}, bottom salinity\cr
[,36] \tab SDO \tab logical \tab cleans up to \code{sdo}, surface dissolved oxygen\cr
[,37] \tab BDO \tab logical \tab cleans up to \code{bdo}, bottom dissolved oxygen\cr
[,38] \tab TEMPAIR \tab numeric \tab cleans up to \code{airtemp}, air temperature\cr
[,39] \tab LATITUDESTART \tab numeric \tab insert_description_here\cr
[,40] \tab LATITUDEEND \tab numeric \tab insert_description_here\cr
[,41] \tab LONGITUDESTART \tab numeric \tab insert_description_here\cr
[,42] \tab LONGITUDEEND \tab numeric \tab insert_description_here\cr
[,43] \tab SPECSTATUSDESCRIPTION \tab character \tab insert_description_here\cr
[,44] \tab LASTUPDATED \tab character \tab insert_description_here\cr
[,45] \tab LIGHTPHASE \tab character \tab insert_description_here\cr
[,46] \tab TIMEZONE \tab character \tab cleans up to \code{timezone}, time zone\cr
[,47] \tab DEPTHSTART \tab integer \tab insert_description_here\cr
[,48] \tab DEPTHEND \tab integer \tab insert_description_here\cr
[,49] \tab PRESSURE \tab numeric \tab insert_description_here\cr
[,50] \tab WINDSPEED \tab integer \tab cleans up to \code{wind}, wind speed (?)\cr
[,51] \tab WINDDIRECTION \tab integer \tab insert_description_here\cr
[,52] \tab WAVEHEIGHT \tab integer \tab insert_description_here\cr
[,53] \tab PRECIPITATION \tab logical \tab insert_description_here\cr
[,54] \tab ESTIMATEDLOC \tab logical \tab insert_description_here\cr
[,55] \tab SEDSIZEDESC \tab character \tab insert_description_here\cr
[,56] \tab BTMCOMPDESC \tab logical \tab insert_description_here\cr
[,57] \tab WEATHERDESC \tab logical \tab insert_description_here\cr
[,58] \tab WATERLVLDESC \tab logical \tab insert_description_here\cr
[,59] \tab ALTERATIONDESC \tab logical \tab insert_description_here\cr
[,60] \tab ACTIVITYDESC \tab logical \tab insert_description_here\cr
[,61] \tab NUMBERREP \tab logical \tab insert_description_here\cr
[,62] \tab COMMENTS \tab logical \tab insert_description_here\cr
}}
\usage{
raw.sa
}
\description{
Raw South Atlantic (Southeast US)  
  
Raw data set for the South Atlantic (Southeast US) bottom trawl survey
}
\keyword{datasets}

