% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.shelf}
\alias{raw.shelf}
\title{Raw Scotian Shelf}
\format{A data.table}
\usage{
raw.shelf
}
\description{
Raw data set for the Scotian Shelf bottom trawl survey
}
\keyword{datasets}

