% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.wcann}
\alias{raw.wcann}
\title{Raw West Coast Annual}
\format{A data.table}
\usage{
raw.wcann
}
\description{
Raw data set for the West Coast Annual bottom trawl survey
}
\keyword{datasets}

