% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw.wctri}
\alias{raw.wctri}
\title{Raw West Coast Triennial}
\format{A data.table}
\usage{
raw.wctri
}
\description{
Raw data set for the West Coast Triennial bottom trawl survey
}
\keyword{datasets}

