% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatValue.R
\name{rmWhite}
\alias{rmWhite}
\title{Remove White Space}
\usage{
rmWhite(x)
}
\arguments{
\item{x}{a data.table}
}
\value{
Nothing, but has the side affect of impacting whatever object was passed as \code{x}.
}
\description{
Remove white space from a data.table
}
\details{
Removes white space from the columns of a data.table that are characters (tested by \code{is.character}). Affects the data.table passed to \code{x}.
}
\seealso{
\code{\link{rm9s}} \code{\link{clean.format}}
}

