% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SODA}
\alias{SODA}
\alias{soda}
\alias{soda.annMax}
\alias{soda.annMean}
\title{SODA
Gridded SODA bottom temperatures}
\format{raster brick}
\usage{
soda

soda.annMax

soda.annMean
}
\description{
From 1958 to 2008. Longitude: 200W to 20E. Latitude: 0N to 90N.
}
\details{
Can be the full data set (\code{soda}), or the annual maximum or mean (\code{soda.annMax} and \code{soda.annMean}, respectively).
}
\keyword{datasets}

