% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spp.corr1}
\alias{spp.corr1}
\title{Corrected Species Names (legacy)
Species names that had been corrected in older version of code}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	3990 obs. of  2 variables:
 $ spp    : chr  "Antipatharian" "Gorgonian" "Gymothorax igromargiatus" "Micropaope uttigii" ...
 $ sppCorr: chr  "Antipatharia" "Gorgonacea" "Gymnothorax nigromargiatus" "Micropanope nuttingi" ...
 - attr(*, ".internal.selfref")=<externalptr> 
}}
\usage{
spp.corr1
}
\description{
Corrected Species Names (legacy)
Species names that had been corrected in older version of code
}
\details{
Used to construct spp.key initially, but should no longer be needed. Provided for reproducibility of older results regarding taxonomy.
Now superceded by \code{\link{getSppData}}. It is recommended that \code{\link{spp.key}} be used for taxonomic reference.
}
\seealso{
\code{\link{getCmmn}} \code{\link{getSpp}} \code{\link{getTax}} \code{\link{spp.key}}
}
\keyword{datasets}

