% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spp.corr1}
\alias{spp.corr1}
\title{Corrected Species Names (legacy)  
  
Species names that had been corrected in older version of code}
\format{A dim = 3990 x 2 data.table data.frame:  
\tabular{rlll}{
[,1] \tab spp \tab character \tab species scientific name; Genus species\cr
[,2] \tab sppCorr \tab character \tab insert_description_here\cr
}}
\usage{
spp.corr1
}
\description{
Corrected Species Names (legacy)  
  
Species names that had been corrected in older version of code
}
\keyword{datasets}

