% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spp.key}
\alias{spp.key}
\title{Species Key
Key to taxonomic and ecological information for all species surveyed
This data.table is the desirable key to use for species names. Relates survey-entered taxonomic identifiers to true scientific names and other related information.
This key was partially derived from the \code{\link{taxInfo}} data set checked by Rachel Berman. However, taxonomy for this data set seems to be an ongoing process, and is always improving. Thus, \code{spp.key} should represent an improvement, and will continue to improve. Check for updates.
To match the values in this key to taxonomic data, one can use \code{merge} using \code{by="ref"}; alternatively, \code{trawlData::match.tbl} may also be convenient for grabbing specific values.}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	6015 obs. of  27 variables:
 $ ref            : chr  "AACATHOBATIDAE" "AACATHOBATIS LOGIROSTRIS" "ACHISTIOIDES ATIGUESIS" "ACTIODOTOIDA" ...
 $ val.src        : chr  NA NA NA NA ...
 $ tbl.row        : int  NA NA NA NA NA NA NA NA NA NA ...
 $ mtch.src       : int  NA NA NA NA NA NA NA NA NA NA ...
 $ tax.src        : chr  NA NA NA NA ...
 $ spp            : chr  NA NA NA NA ...
 $ common         : chr  NA NA NA NA ...
 $ taxLvl         : chr  NA NA NA NA ...
 $ species        : chr  NA NA NA NA ...
 $ genus          : chr  NA NA NA NA ...
 $ family         : chr  NA NA NA NA ...
 $ order          : chr  NA NA NA NA ...
 $ class          : chr  NA NA NA NA ...
 $ superclass     : chr  NA NA NA NA ...
 $ subphylum      : chr  NA NA NA NA ...
 $ phylum         : chr  NA NA NA NA ...
 $ kingdom        : chr  NA NA NA NA ...
 $ trophicDiet    : chr  NA NA NA NA ...
 $ trophicOrig    : chr  NA NA NA NA ...
 $ Picture        : chr  NA NA NA NA ...
 $ trophicLevel   : num  NA NA NA NA NA NA NA NA NA NA ...
 $ trophicLevel.se: num  NA NA NA NA NA NA NA NA NA NA ...
 $ tax.src2       : chr  NA NA NA NA ...
 $ conflict       : logi  NA NA NA NA NA NA ...
 $ flag           : chr  "ok" "ok" "ok" "ok" ...
 $ website        : chr  NA NA NA NA ...
 $ website2       : chr  NA NA NA NA ...
 - attr(*, ".internal.selfref")=<externalptr> 
 - attr(*, "sorted")= chr  "spp" "ref"
}}
\usage{
spp.key
}
\description{
Species Key
Key to taxonomic and ecological information for all species surveyed
This data.table is the desirable key to use for species names. Relates survey-entered taxonomic identifiers to true scientific names and other related information.
This key was partially derived from the \code{\link{taxInfo}} data set checked by Rachel Berman. However, taxonomy for this data set seems to be an ongoing process, and is always improving. Thus, \code{spp.key} should represent an improvement, and will continue to improve. Check for updates.
To match the values in this key to taxonomic data, one can use \code{merge} using \code{by="ref"}; alternatively, \code{trawlData::match.tbl} may also be convenient for grabbing specific values.
}
\keyword{datasets}

