% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spp.key}
\alias{spp.key}
\title{Species Key
Key to taxonomic and ecological information for all species surveyed
This data.table is the desirable key to use for species names. Relates survey-entered taxonomic identifiers to true scientific names and other related information.
This key was partially derived from the \code{\link{taxInfo}} data set checked by Rachel Berman. However, taxonomy for this data set seems to be an ongoing process, and is always improving. Thus, \code{spp.key} should represent an improvement, and will continue to improve. Check for updates.
To match the values in this key to taxonomic data, one can use \code{merge} using \code{by="ref"}; alternatively, \code{trawlData::match.tbl} may also be convenient for grabbing specific values.}
\format{A dim = 6015 x 27 data.table data.frame:
\tabular{rlll}{
[,1] \tab ref \tab character \tab insert_description_here\cr
[,2] \tab val.src \tab character \tab insert_description_here\cr
[,3] \tab tbl.row \tab integer \tab insert_description_here\cr
[,4] \tab mtch.src \tab integer \tab insert_description_here\cr
[,5] \tab tax.src \tab character \tab insert_description_here\cr
[,6] \tab spp \tab character \tab insert_description_here\cr
[,7] \tab common \tab character \tab insert_description_here\cr
[,8] \tab taxLvl \tab character \tab insert_description_here\cr
[,9] \tab species \tab character \tab insert_description_here\cr
[,10] \tab genus \tab character \tab insert_description_here\cr
[,11] \tab family \tab character \tab insert_description_here\cr
[,12] \tab order \tab character \tab insert_description_here\cr
[,13] \tab class \tab character \tab insert_description_here\cr
[,14] \tab superclass \tab character \tab insert_description_here\cr
[,15] \tab subphylum \tab character \tab insert_description_here\cr
[,16] \tab phylum \tab character \tab insert_description_here\cr
[,17] \tab kingdom \tab character \tab insert_description_here\cr
[,18] \tab trophicDiet \tab character \tab insert_description_here\cr
[,19] \tab trophicOrig \tab character \tab insert_description_here\cr
[,20] \tab Picture \tab character \tab insert_description_here\cr
[,21] \tab trophicLevel \tab numeric \tab insert_description_here\cr
[,22] \tab trophicLevel.se \tab numeric \tab insert_description_here\cr
[,23] \tab tax.src2 \tab character \tab insert_description_here\cr
[,24] \tab conflict \tab logical \tab insert_description_here\cr
[,25] \tab flag \tab character \tab insert_description_here\cr
[,26] \tab website \tab character \tab insert_description_here\cr
[,27] \tab website2 \tab character \tab insert_description_here\cr
}}
\usage{
spp.key
}
\description{
Species Key
Key to taxonomic and ecological information for all species surveyed
This data.table is the desirable key to use for species names. Relates survey-entered taxonomic identifiers to true scientific names and other related information.
This key was partially derived from the \code{\link{taxInfo}} data set checked by Rachel Berman. However, taxonomy for this data set seems to be an ongoing process, and is always improving. Thus, \code{spp.key} should represent an improvement, and will continue to improve. Check for updates.
To match the values in this key to taxonomic data, one can use \code{merge} using \code{by="ref"}; alternatively, \code{trawlData::match.tbl} may also be convenient for grabbing specific values.
}
\keyword{datasets}

