% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperStat.R
\name{sumna}
\alias{sumna}
\title{Sum NA}
\usage{
sumna(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{numbers to be summed}

\item{na.rm}{logical, default TRUE, remove NA's?}

\item{...}{unused}
}
\description{
Sum where NA's are removed, but if all are NA, return NA (not 0)
}
\seealso{
\code{\link{sumna}}, \code{\link{meanna}}, \code{\link{wtAvg}}, \code{\link{fill.mean}}
}

