% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{taxInfo}
\alias{taxInfo}
\title{Taxonomic Information (legacy)
Taxonomic classification and ecological information
This is the data set manually checked by Rachel Berman.
Superceded by \code{\link{spp.key}}.}
\format{\preformatted{Classes ‘data.table’ and 'data.frame':	4483 obs. of  20 variables:
 $ raw.spp        : chr  "Abietinaria sp." "Abietinaria sp. A (Clark 2006)" "Abietinaria greenei" "ABISA" ...
 $ spp            : chr  "Abietinaria" "Abietinaria" "Abietinaria greenei" "Abisa" ...
 $ common         : chr  "sea fir" "sea fir" "sea fir" "bivalve" ...
 $ taxLvl         : chr  "Genus" "Genus" "Species" NA ...
 $ species        : chr  NA NA "Abietinaria greenei" NA ...
 $ genus          : chr  "Abietinaria" "Abietinaria" "Abietinaria" NA ...
 $ family         : chr  "Sertulariidae" "Sertulariidae" "Sertulariidae" NA ...
 $ order          : chr  "Leptothecatae" "Leptothecatae" "Leptothecatae" NA ...
 $ class          : chr  "Hydrozoa" "Hydrozoa" "Hydrozoa" NA ...
 $ superclass     : chr  NA NA NA NA ...
 $ subphylum      : chr  "Medusozoa" "Medusozoa" "Medusozoa" NA ...
 $ phylum         : chr  "Cnidaria" "Cnidaria" "Cnidaria" NA ...
 $ kingdom        : chr  "Animalia" "Animalia" "Animalia" NA ...
 $ isSpecies      : logi  FALSE FALSE TRUE FALSE TRUE TRUE ...
 $ correctSpp     : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
 $ trophicDiet    : chr  NA NA NA NA ...
 $ trophicOrig    : chr  NA NA NA NA ...
 $ Picture        : chr  "y" "y" "n" "n" ...
 $ trophicLevel   : num  NA NA NA NA 4.5 NA NA NA NA NA ...
 $ trophicLevel.se: num  NA NA NA NA 0.8 NA NA NA NA NA ...
 - attr(*, "sorted")= chr "spp"
 - attr(*, ".internal.selfref")=<externalptr> 
 - attr(*, "index")= atomic  
  ..- attr(*, "__raw.spp")= int  184 185 186 187 188 189 197 198 196 4 ...
}}
\usage{
taxInfo
}
\description{
Taxonomic Information (legacy)
Taxonomic classification and ecological information
This is the data set manually checked by Rachel Berman.
Superceded by \code{\link{spp.key}}.
}
\keyword{datasets}

