% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{taxInfo}
\alias{taxInfo}
\title{Taxonomic Information (legacy)
Taxonomic classification and ecological information
This is the data set manually checked by Rachel Berman.
Superceded by \code{\link{spp.key}}.}
\format{A dim = 4483 x 20 data.table data.frame:
\tabular{rlll}{
[,1] \tab raw.spp \tab character \tab insert_description_here\cr
[,2] \tab spp \tab character \tab insert_description_here\cr
[,3] \tab common \tab character \tab insert_description_here\cr
[,4] \tab taxLvl \tab character \tab insert_description_here\cr
[,5] \tab species \tab character \tab insert_description_here\cr
[,6] \tab genus \tab character \tab insert_description_here\cr
[,7] \tab family \tab character \tab insert_description_here\cr
[,8] \tab order \tab character \tab insert_description_here\cr
[,9] \tab class \tab character \tab insert_description_here\cr
[,10] \tab superclass \tab character \tab insert_description_here\cr
[,11] \tab subphylum \tab character \tab insert_description_here\cr
[,12] \tab phylum \tab character \tab insert_description_here\cr
[,13] \tab kingdom \tab character \tab insert_description_here\cr
[,14] \tab isSpecies \tab logical \tab insert_description_here\cr
[,15] \tab correctSpp \tab logical \tab insert_description_here\cr
[,16] \tab trophicDiet \tab character \tab insert_description_here\cr
[,17] \tab trophicOrig \tab character \tab insert_description_here\cr
[,18] \tab Picture \tab character \tab insert_description_here\cr
[,19] \tab trophicLevel \tab numeric \tab insert_description_here\cr
[,20] \tab trophicLevel.se \tab numeric \tab insert_description_here\cr
}}
\usage{
taxInfo
}
\description{
Taxonomic Information (legacy)
Taxonomic classification and ecological information
This is the data set manually checked by Rachel Berman.
Superceded by \code{\link{spp.key}}.
}
\keyword{datasets}

